rename the package
