+++ src/main/java/org/apache/camel/idea/completion/extension/CamelPropertiesSmartCompletionExtension.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.idea.completionproviders;
+package org.apache.camel.idea.completion.extension;
 
 import com.intellij.codeInsight.completion.CompletionParameters;
 import com.intellij.codeInsight.completion.CompletionResultSet;
